pyproject.toml
[tool.poetry]
name = "fastapi-supabase"
version = "0.1.0"
description = "Dependencies of project"
authors = ["RuanCarlosSilva <143547755+ruansilva123@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
supabase = "^2.9.1"
python-decouple = "^3.8"
pydantic = "^2.9.2"
pyjwt = "^2.9.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
